Remove code doing a no-op due to float -> int rounding. Make implicit float -> int/long conversions explicit.
(Implicit float -> int conversions can be found by compiling with
-Wconversion in gcc [versions 4.1.1 and 4.2.4, and surely many others].)

Landing the patch for Jacob Mandelson, original review: http://codereview.chromium.org/201091

BUG=none
TEST=app_unittests & base_unittests

Review URL: http://codereview.chromium.org/200122

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26119 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 655750cba246a2e3049867c3b200ca478f883230

+++ chrome/browser/gtk/back_forward_button_gtk.cc
@@ -110,7 +110,7 @@ gboolean BackForwardButtonGtk::OnButtonPress(GtkWidget* widget,
   if (event->button != 1)
     return FALSE;
 
-  button->y_position_of_last_press_ = event->y;
+  button->y_position_of_last_press_ = static_cast<int>(event->y);
   MessageLoop::current()->PostDelayedTask(FROM_HERE,
       button->show_menu_factory_.NewRunnableMethod(
           &BackForwardButtonGtk::ShowBackForwardMenu),
